refactor: Change types inside formSelectInput
